requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15.0
drf-nested-routers>=0.93.5
django-cors-headers>=4.3.0
django-filter>=24.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication
djangorestframework-simplejwt>=5.3.1

# iCal
icalendar>=5.0.0
recurring-ical-events>=2.2.0
requests>=2.31.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Celery (for background tasks)
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Development
ipython>=8.0.0
